optimize: float division in ADC filter, IWDG watchdog, frequency EMA
- Adc_Driver: Filter_To_Voltage integer→float division, eliminates LSB truncation
- Main: IWDG independent watchdog (LSI 40kHz, 1.6s timeout), feed in main loop
- Ui_Controller: frequency display EMA smoothing (α=0.25), same as V/I

+++ Keil_Project/Hardware/Adc_Driver.c
@@ -44,7 +44,7 @@ static void Filter_Push(Filter_Window* fw, uint16_t new_val)
 
 static float Filter_To_Voltage(const Filter_Window* fw)
 {
-    return ((float)(fw->accum / fw->filled) / 4095.0f) * VREF_MCU;
+    return ((float)fw->accum / (float)fw->filled / 4095.0f) * VREF_MCU;
 }
 
 /* ── 模块状态 ── */
